refactor(utils): 共通例外クラスをutils/exceptions.pyに抽出
Issue #460に基づき、APIクライアント共通の例外階層とASYNC_FATAL_EXCEPTIONSを
utils/exceptions.pyに集約し、api_client.py/github_client.pyからの重複を除去。

- 新規: utils/exceptions.py (APIClientError, APIConnectionError, APITimeoutError,
  APIHTTPError, APIRetryError, APIJSONDecodeError, ASYNC_FATAL_EXCEPTIONS)
- 変更: utils/api_client.py - 例外定義を削除、utils.exceptionsからimport
- 変更: utils/github_client.py - ASYNC_FATAL_EXCEPTIONSをutils.exceptionsからimport
- テスト: test_github_client.py - 共有例外階層の互換性を検証

全1381テスト通過 (カバレッジ96.17%)、ruff/mypyクリーン。

Refs: #460
