+++ utils/api_client.py
@@ -14,18 +14,17 @@ from structlog.typing import FilteringBoundLogger
 
 from config.settings import settings
 from models.responses import Album, Comment, Photo, Post, Todo, User
+from utils.exceptions import (
+    ASYNC_FATAL_EXCEPTIONS,
+    APIClientError,
+    APIConnectionError,
+    APIHTTPError,
+    APIJSONDecodeError,
+    APIRetryError,
+    APITimeoutError,
+)
 from utils.logger import get_logger
 
-# httpx 例外（NetworkError, RemoteProtocolError 等）をここに追加してはならない。
-# github_client.py の _request では except 句の順序で先行キャッチしており、
-# この定数に httpx 例外を追加すると NetworkError のリトライが壊れる。
-ASYNC_FATAL_EXCEPTIONS: tuple[type[BaseException], ...] = (
-    KeyboardInterrupt,
-    SystemExit,
-    MemoryError,
-    RecursionError,
-    asyncio.CancelledError,
-)
 SYNC_FATAL_EXCEPTIONS: tuple[type[BaseException], ...] = (
     KeyboardInterrupt,
     SystemExit,
@@ -74,44 +73,6 @@ def _validate_optional_int(value: int | None, name: str, min_value: int) -> None
     """
     if value is not None and value < min_value:
         raise ValueError(f"{name} must be >= {min_value}")
-
-
-# =============================================================================
-# 例外クラス
-# =============================================================================
-
-
-class APIClientError(Exception):
-    """APIクライアント基底例外"""
-
-
-class APIConnectionError(APIClientError):
-    """API接続エラー"""
-
-
-class APITimeoutError(APIClientError):
-    """APIタイムアウトエラー"""
-
-
-class APIHTTPError(APIClientError):
-    """HTTPステータスエラー"""
-
-    def __init__(self, message: str, status_code: int, response: httpx.Response | None = None):
-        super().__init__(message)
-        self.status_code = status_code
-        self.response = response
-
-
-class APIRetryError(APIClientError):
-    """リトライ上限エラー"""
-
-
-class APIJSONDecodeError(APIClientError):
-    """JSONパース＋スキーマ検証エラー"""
-
-    def __init__(self, message: str, response: httpx.Response | None = None):
-        super().__init__(message)
-        self.response = response
 
 
 def _safe_parse_json(response: httpx.Response) -> Any:
